remove old dataloader
